requirements.txt
scikit-learn
numpy
pandas
ultralytics
pyyaml